+++ JRPG/Game.h
@@ -17,14 +17,14 @@ public:
 	sf::Vector2i GetMapSize() const;
 	sf::Vector2f GetPlayerWorldPosition() const;
 	int GetNbLayer() const;
-	sf::Vector2f GetTileSize() const;
+	sf::Vector2i GetTileSize() const;
 	int GetLayerCell(int,int,int) const;
 	std::string GetLayerName(int) const;
 
 	void SetTile(Tile);
 	void SetPlayerWorldPosition(sf::Vector2f);
 	void SetMapSize(sf::Vector2i);
-	void SetTileSize(sf::Vector2f);
+	void SetTileSize(sf::Vector2i);
 	void SetLayer(std::string, std::string);
 
 	//movement
